Validate and Cleaning up

+++ products/cart.php
@@ -53,50 +53,55 @@ if (isset($_POST['checkout'])) {
 		<div class="row">
 			<div class="col-md-12 ftco-animate">
 				<div class="cart-list">
-					<table class="table">
-						<thead class="thead-primary">
-							<tr class="text-center">
-								<th>&nbsp;</th>
-								<th>&nbsp;</th>
-								<th>Product</th>
-								<th>Price</th>
-								<th>Quantity</th>
-								<th>Total</th>
-							</tr>
-						</thead>
-						<tbody>
-							<?php foreach ($allproducts as $product): ?>
+					<?php if (count($allproducts) > 0): ?>
+
+						<table class="table">
+							<thead class="thead-primary">
 								<tr class="text-center">
-									<td class="product-remove"><a
-											href="delete-product.php?id=<?php echo $product->id; ?>"><span
-												class="icon-close"></span></a></td>
-
-									<td class="image-prod">
-										<div class="img"
-											style="background-image:url(<?php echo APPURL; ?>/images/<?php echo $product->image; ?>);">
-										</div>
-									</td>
-
-									<td class="product-name">
-										<h3><?php echo $product->name; ?></h3>
-										<p><?php echo $product->description; ?></p>
-									</td>
-
-									<td class="price">$<?php echo $product->price; ?></td>
-
-									<td>
-										<div class="input-group mb-3">
-											<input disabled type="text" name="quantity"
-												class="quantity form-control input-number"
-												value="<?php echo $product->quantity; ?>" min="1" max="100">
-										</div>
-									</td>
-
-									<td class="total">$<?php echo $product->price * $product->quantity; ?></td>
+									<th>&nbsp;</th>
+									<th>&nbsp;</th>
+									<th>Product</th>
+									<th>Price</th>
+									<th>Quantity</th>
+									<th>Total</th>
 								</tr>
-							<?php endforeach; ?>
-						</tbody>
-					</table>
+							</thead>
+							<tbody>
+								<?php foreach ($allproducts as $product): ?>
+									<tr class="text-center">
+										<td class="product-remove"><a
+												href="delete-product.php?id=<?php echo $product->id; ?>"><span
+													class="icon-close"></span></a></td>
+
+										<td class="image-prod">
+											<div class="img"
+												style="background-image:url(<?php echo APPURL; ?>/images/<?php echo $product->image; ?>);">
+											</div>
+										</td>
+
+										<td class="product-name">
+											<h3><?php echo $product->name; ?></h3>
+											<p><?php echo $product->description; ?></p>
+										</td>
+
+										<td class="price">$<?php echo $product->price; ?></td>
+
+										<td>
+											<div class="input-group mb-3">
+												<input disabled type="text" name="quantity"
+													class="quantity form-control input-number"
+													value="<?php echo $product->quantity; ?>" min="1" max="100">
+											</div>
+										</td>
+
+										<td class="total">$<?php echo $product->price * $product->quantity; ?></td>
+									</tr>
+								<?php endforeach; ?>
+							</tbody>
+						</table>
+					<?php else: ?>
+						<p>Your cart is empty, add products</p>
+					<?php endif; ?>
 				</div>
 			</div>
 		</div>
@@ -119,15 +124,17 @@ if (isset($_POST['checkout'])) {
 					<hr>
 					<p class="d-flex total-price">
 						<span>Total</span>
-						<span>$<?php echo $allcart_total->total + 10 - 3; ?> </span>
-						<input type="hidden" value="<?php echo $allcart_total->total + 10 - 3; ?>">
+						<?php if ($allcart_total->total > 0): ?>
+							<span>$<?php echo $allcart_total->total + 10 - 3; ?> </span>
+							<input type="hidden" value="<?php echo $allcart_total->total + 10 - 3; ?>">
+						<?php endif; ?>
 					</p>
 				</div>
 				<form method="POST" action="cart.php">
 					<input type="hidden" name="total_price" value="<?php echo $allcart_total->total + 10 - 3; ?>">
-
-					<button name="checkout" type="submit" class="btn btn-primary py-3 px-4">Proceed to Checkout</button>
-
+					<?php if ($allcart_total->total > 0): ?>
+						<button name="checkout" type="submit" class="btn btn-primary py-3 px-4">Proceed to Checkout</button>
+					<?php endif; ?>
 				</form>
 			</div>
 		</div>
